Make ems_network_controller_spec a shared spec
Make ems_network_controller_spec a shared spec, which can be
called from a provider gem as with a provider parameter.
